refactor(ssd1306.h): add a function to manage SSD1306 write operations
SSD1306 supports several communication interfaces such as I2C, SPI or parallel.
In order to keep the code reusable and port it easily, it is required to add a
function that manages the write operations by selecting the right write function
defined in the ssd1306_t struct based on the specified interface.

- add _ssd1306_write function
- modify ssd1306_t struct to have a dedicated function to write using the I2C interface
- add a enum to list SSD1306 control bytes
- add missing ACTIVATE_SCROLL command in the ssd1306_activate_scroll function